tests/wpt/tests/css/css-grid/grid-lanes/tentative/track-sizing/auto-repeat/row-auto-repeat-010.html

<!DOCTYPE html>
<title>Auto repeat tracks and percentage max sizes</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<meta name="assert" content="This test checks that auto repeat tracks use percentage max size when size is indefinite to compute the number of tracks.">
<link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
<style>
.grid-lanes {
    position: relative;
    display: grid-lanes;
    masonry-direction: row;
    grid-template-rows: repeat(auto-fill, 50px);
    max-width: 50%;
    max-height: 80%;
    background: pink;
}
.wrapper {
  width: 600px;
  height: 250px;
}
.item {
  background: lime;
  /* Place item on the last row. */
  grid-row: -2;
  width: 300px;
  height: 100%;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.grid-lanes');">

<div id="log"></div>

<div class="wrapper">
  <div class="grid-lanes" data-expected-width="300" data-expected-height="200">
    <div class="item" data-offset-x="0" data-offset-y="150" data-expected-width="300" data-expected-height="50"></div>
  </div>
</div>

<div class="wrapper">
  <div class="grid-lanes" style="width: 200px; height: 100px;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-offset-x="0" data-offset-y="50" data-expected-width="300" data-expected-height="50"></div>
  </div>
</div>

<div class="wrapper">
  <div class="grid-lanes" style="width: min-content; height: min-content;" data-expected-width="300" data-expected-height="200">
    <div class="item" data-offset-x="0" data-offset-y="150" data-expected-width="300" data-expected-height="50"></div>
  </div>
</div>

<div class="wrapper">
  <div class="grid-lanes" style="width: max-content; height: max-content;" data-expected-width="300" data-expected-height="200">
    <div class="item" data-offset-x="0" data-offset-y="150" data-expected-width="300" data-expected-height="50"></div>
  </div>
</div>

</body>
