third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/row-auto-repeat-010-expected.txt

This is a testharness.js-based test.
[FAIL] .masonry 1
  assert_equals: \n<div class="masonry" data-expected-width="300" data-expected-height="200">\n    <div class="item" data-offset-x="0" data-offset-y="150" data-expected-width="300" data-expected-height="50"></div>\n  </div>\nheight expected 200 but got 0
[FAIL] .masonry 2
  assert_equals: \n<div class="masonry" style="width: 200px; height: 100px;" data-expected-width="200" data-expected-height="100">\n    <div class="item" data-offset-x="0" data-offset-y="50" data-expected-width="300" data-expected-height="50"></div>\n  </div>\nheight expected 50 but got 100
[FAIL] .masonry 3
  assert_equals: \n<div class="masonry" style="width: min-content; height: min-content;" data-expected-width="300" data-expected-height="200">\n    <div class="item" data-offset-x="0" data-offset-y="150" data-expected-width="300" data-expected-height="50"></div>\n  </div>\nheight expected 200 but got 0
[FAIL] .masonry 4
  assert_equals: \n<div class="masonry" style="width: max-content; height: max-content;" data-expected-width="300" data-expected-height="200">\n    <div class="item" data-offset-x="0" data-offset-y="150" data-expected-width="300" data-expected-height="50"></div>\n  </div>\nheight expected 200 but got 0
Harness: the test ran to completion.
